Main Menu / Text Package
Completed main menu. Process included the addition of a free text package now included in the assests. Menu is simple with only options to start a new game or quit the game. New game option causes the "main" scene to appear as though the game has begun. Also added an attempt at a build of the project now that two scenes are included.